@@ -1,6 +1,6 @@
 package com.github.phoswald.sample.sample;
 
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement(name = "EchoResponse")
 public class EchoResponse {
